EVG-19927: Fix linter errors.
This required refactoring some of oom_tracker.go into a helper
file that builds only on darwin/Linux.
